requirements.txt
discord.py
tmdbsimple
influxdb